tools/processor/pyproject.toml
[project]
name = "processor"
version = "1"
requires-python="==3.12.*"
dependencies = [
    "psycopg2-binary",
    "PyYAML",
    "dotenv"
]